Place parameters and position group in core.py
