+++ src/spyglass/decoding/v1/clusterless.py
@@ -15,15 +15,14 @@ import datajoint as dj
 import numpy as np
 import pandas as pd
 import xarray as xr
-from non_local_detector.models import ContFragClusterlessClassifier
 from non_local_detector.models.base import ClusterlessDetector
 
 from spyglass.common.common_interval import IntervalList  # noqa: F401
 from spyglass.common.common_position import IntervalPositionInfo
-from spyglass.decoding.v1.dj_decoder_conversion import (
-    convert_classes_to_dict,
-    restore_classes,
-)
+from spyglass.decoding.v1.core import (
+    DecodingParameters,
+    PositionGroup,
+)  # noqa: F401
 from spyglass.decoding.v1.waveform_features import (
     UnitWaveformFeatures,
 )  # noqa: F401
@@ -32,39 +31,6 @@ from spyglass.settings import config
 from spyglass.utils import SpyglassMixin, logger
 
 schema = dj.schema("decoding_clusterless_v1")
-
-
-@schema
-class DecodingParameters(SpyglassMixin, dj.Lookup):
-    """Parameters for decoding the animal's mental position and some category of interest"""
-
-    definition = """
-    decoding_param_name : varchar(80)  # a name for this set of parameters
-    ---
-    decoding_params : BLOB             # initialization parameters for model
-    decoding_kwargs : BLOB             # additional keyword arguments
-    """
-
-    # contents = [
-    #     [
-    #         "contfrag_clusterless",
-    #         vars(ContFragClusterlessClassifier()),
-    #         dict(),
-    #     ],
-    # ]
-
-    # @classmethod
-    # def insert_default(cls):
-    #     cls.insert(cls.contents, skip_duplicates=True)
-
-    # def insert(self, keys, **kwargs):
-    #     pass
-
-    def insert1(self, key, **kwargs):
-        super().insert1(convert_classes_to_dict(key), **kwargs)
-
-    def fetch1(self, *args, **kwargs):
-        return restore_classes(super().fetch1(*args, **kwargs))
 
 
 @schema
@@ -88,43 +54,6 @@ class UnitWaveformFeaturesGroup(SpyglassMixin, dj.Manual):
                 {
                     **key,
                     "waveform_features_group_name": group_name,
-                },
-                skip_duplicates=True,
-            )
-
-
-@schema
-class PositionGroup(SpyglassMixin, dj.Manual):
-    definition = """
-    position_group_name: varchar(80)
-    ----
-    position_variables = NULL: longblob # list of position variables to decode
-    """
-
-    class Position(SpyglassMixin, dj.Part):
-        definition = """
-        -> PositionGroup
-        -> PositionOutput.proj(pos_merge_id='merge_id')
-        """
-
-    def create_group(
-        self,
-        group_name: str,
-        keys: list[dict],
-        position_variables: list[str] = ["position_x", "position_y"],
-    ):
-        self.insert1(
-            {
-                "position_group_name": group_name,
-                "position_variables": position_variables,
-            },
-            skip_duplicates=True,
-        )
-        for key in keys:
-            self.Position.insert1(
-                {
-                    **key,
-                    "position_group_name": group_name,
                 },
                 skip_duplicates=True,
             )
